GAM update
After mis-interpretation, changing GAMM to GAM in script #3.  Causes
cascading changes to be made in 3.1, 3.2, and 3.4.  The same with
script series #4.  Updated script7 with new climate years, and created
marker year growth deviations in script #9.  Script #9 also contains
ANOVA’s for determining differences in the weights and BAI for ambient
growth and extreme year growth.